+++ .circleci/config.yml
@@ -74,7 +74,6 @@ jobs:
       - checkout
       - *attach_workspace
       - run: npm run dist
-      - run: npm run test -- ./scripts/check-theme-vars.test.js
       - run: node ./tests/dekko/dist.test.js
       - run: npm run bundlesize
       - persist_to_workspace:
